Implement Lakka core
 * Rebased on LibreELEC master
 * Based on Lakka LE9.1 branch
 * Added Vulkan support for AMD and Intel
 * Removed all cores
 * Minimal changes on LibreELEC based packets
 * Added core dependency meta package
 * Added libgomp library for cores

This distribution should be used as a small platform to run
the libretro cores that are build by the buildbot.

We still need to activate the core update in retroarch.

If some targets need special builds that the buildbot is not
providing, we better update the buildbot to provide them.

That way we create an easy to maintain base distribution.